+++ lib/pronto/formatter/bitbucket_pull_request_formatter.rb
@@ -8,6 +8,10 @@ module Pronto
       def pretty_name
         'BitBucket'
       end
+
+      def line_number(message, _)
+        message.line.line.new_lineno if message.line
+      end
     end
   end
 end